ehr-main/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-storages==1.13.2
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
